heydev-backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
opencv-python
whisper
